Rename ssl_debug_helpers.h
It's no longer generated, so rename it accordingly.
