+++ library/ssl_tls13_server.c
@@ -26,7 +26,7 @@
 #include "mbedtls/debug.h"
 
 #include "ssl_misc.h"
-#include "ssl_debug_helpers_generated.h"
+#include "ssl_debug_helpers.h"
 
 int mbedtls_ssl_tls13_handshake_server_step( mbedtls_ssl_context *ssl )
 {
